refactor: remove gRPC bridge, port SELL post-processing to reconcileFills
Claude + Codex のピアレビュー合意に基づく cleanup (#106 塩漬けの物理削除)。

## bridge が無くても動くようにしてから削除
- reconcileFills.applyFillToState に TradeEventHandler から以下を port
  - SELL FILLED で realized_pnl < 0 のとき setCooldown
    (pullbackScheduler が state.cooldownUntil 読んで entry を抑止する経路、
     port しないと負けトレード後も即 re-entry する退行)
  - 全 terminal status (FILLED / CANCELLED / REJECTED / EXPIRED) で
    clearPendingOrder。pullbackScheduler が lockPendingOrder してるので
    release しないと symbol が永久ロック気味になる
- 意図的に port しなかった機能
  - addPendingSettlement (US T+2 cash tracking) — POC は settled_cash
    gate を cron で使っていない。必要になったら issue で戻す

## 削除
- bridge/ ディレクトリ全体 (Node gRPC service, Dockerfile)
- src/trading/bridge/ (BridgeContainer DO, keepAlive, schedule)
- src/trading/events/ (TradeEventHandler) + src/trading/application/TradeEventService
- src/routes/events.ts (/events/trade ingest) + src/infrastructure/webull/TradeEventBridge
- src/trading/domain/TradeEvent
- wrangler.jsonc から BRIDGE binding / containers[] を削除 + migration
  v4 `deleted_classes: ["BridgeContainer"]` を追加
- Env から BRIDGE / WEBULL_GRPC_ENDPOINT / EVENT_INGEST_URL /
  EVENT_INGEST_SECRET を削除
- package.json から @cloudflare/containers 依存 削除
- .github/workflows/ci.yml の bridge job 削除
- .coderabbit.yaml の bridge/* path_instructions 削除
- test/events/, test/trading/bridge/, test/routes/events.test.ts,
  test/helpers/configFixtures.ts の bridgeRunMode 削除

## 2 段階 rollout — 今回は Phase 1
Code が `global_config.bridge_run_mode` 列を読まなくなる状態にして
deploy。D1 の列 DROP は次 PR (Phase 2)。Codex 指摘通り同 deploy で
schema と code を同時変更しない方が rollback 窓を確保できる。

## Secret cleanup
`WEBULL_GRPC_ENDPOINT` / `EVENT_INGEST_URL` / `EVENT_INGEST_SECRET` は
deploy 確認後に手動 `wrangler secret delete` 推奨。

refs #106
