+++ .github/workflows/ci.yml
@@ -36,30 +36,3 @@ jobs:
 
       - name: Test (worker)
         run: pnpm test
-
-  bridge:
-    name: bridge (typecheck + test)
-    runs-on: ubuntu-latest
-    timeout-minutes: 10
-    defaults:
-      run:
-        working-directory: bridge
-    steps:
-      - uses: actions/checkout@v4
-
-      - uses: pnpm/action-setup@v4
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 24
-          cache: pnpm
-          cache-dependency-path: bridge/pnpm-lock.yaml
-
-      - name: Install bridge deps
-        run: pnpm install --frozen-lockfile
-
-      - name: Typecheck
-        run: pnpm run typecheck
-
-      - name: Test
-        run: pnpm test
