--- bridge/.env.example
@@ -1,6 +1,0 @@
-WEBULL_APP_KEY=
-WEBULL_APP_SECRET=
-WEBULL_ACCOUNT_ID_JP_CASH=
-WEBULL_GRPC_ENDPOINT=events-api.sandbox.webull.hk:443
-EVENT_INGEST_URL=https://example.workers.dev/events/trade
-EVENT_INGEST_SECRET=
