--- bridge/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "webull-trading-bridge",
-  "private": true,
-  "type": "module",
-  "packageManager": "pnpm@10.33.0",
-  "scripts": {
-    "start": "tsx grpc/index.ts",
-    "typecheck": "tsc --noEmit -p .",
-    "test": "vitest run"
-  },
-  "dependencies": {
-    "@grpc/grpc-js": "^1.14.0",
-    "@grpc/proto-loader": "^0.8.0"
-  },
-  "devDependencies": {
-    "@types/node": "^24.6.0",
-    "tsx": "^4.20.6",
-    "typescript": "^5.6.3",
-    "vitest": "^2.1.9"
-  }
-}
