+++ package.json
@@ -14,7 +14,6 @@
     "accounts": "tsx scripts/list-webull-accounts.ts"
   },
   "dependencies": {
-    "@cloudflare/containers": "^0.3.2",
     "drizzle-orm": "^0.45.2",
     "hono": "^4.12.14"
   },
