+++ src/app.ts
@@ -3,10 +3,8 @@ import { HTTPException } from 'hono/http-exception'
 import type { Env } from './config/env'
 import { auditLogger } from './infrastructure/logger/AuditLogger'
 import { basicAuthMiddleware } from './middleware/basicAuth'
-import { TRADE_EVENT_INGEST_SECRET_HEADER } from './infrastructure/webull/TradeEventBridge'
 import { health } from './routes/health'
 import { trade } from './routes/trade'
-import { events } from './routes/events'
 // Webull routes (Phase 2 append)
 import { webull } from './routes/webull'
 import { BrokerRequestError, TradingError, ValidationError } from './shared/errors'
@@ -48,19 +46,8 @@ export function createApp() {
   const app = new Hono<AppBindings>()
   app.use('*', auditLogger())
   app.use('/trade/*', basicAuthMiddleware())
-  app.use('/events/*', async (c, next) => {
-    const secret = c.env.EVENT_INGEST_SECRET
-    const providedSecret = c.req.header(TRADE_EVENT_INGEST_SECRET_HEADER)
-
-    if (!secret || !providedSecret || !timingSafeEqual(providedSecret, secret)) {
-      return c.json({ error: 'Unauthorized' }, 401)
-    }
-
-    await next()
-  })
   app.route('/health', health)
   app.route('/trade', trade)
-  app.route('/events', events)
   // Webull routes (Phase 2 append)
   app.use('/webull/*', basicAuthMiddleware())
   app.route('/webull', webull)
@@ -70,20 +57,4 @@ export function createApp() {
   return app
 }
 
-function timingSafeEqual(a: string, b: string) {
-  const enc = new TextEncoder()
-  const ab = enc.encode(a)
-  const bb = enc.encode(b)
-
-  if (ab.byteLength !== bb.byteLength) {
-    return false
-  }
-
-  let diff = 0
-  for (let i = 0; i < ab.byteLength; i++) {
-    diff |= ab[i]! ^ bb[i]!
-  }
-
-  return diff === 0
-}
 import { admin } from './routes/admin'
