--- src/infrastructure/webull/TradeEventBridge.ts
@@ -1,7 +1,0 @@
-import type { TradeEvent } from '../../trading/domain/TradeEvent'
-
-export const TRADE_EVENT_INGEST_SECRET_HEADER = 'x-event-ingest-secret'
-
-export interface TradeEventIngestRequest {
-  event: TradeEvent
-}
