+++ src/trading/state/PortfolioStore.ts
@@ -2,7 +2,7 @@ import type { PortfolioState } from './portfolioTypes'
 
 /**
  * The subset of {@link PortfolioStateDO} that TradingService and
- * TradeEventHandler need. Kept independent from {@link PositionStore} on
+ * reconcileFills need. Kept independent from {@link PositionStore} on
  * purpose — portfolio-level state is not per-symbol.
  */
 export interface PortfolioStore {
