+++ test/routes/trade.test.ts
@@ -14,7 +14,6 @@ vi.mock('../../src/infrastructure/db/symbolUniverse', () => ({
 const env = {
   BASIC_AUTH_USER: 'admin',
   BASIC_AUTH_PASSWORD: 'secret',
-  EVENT_INGEST_SECRET: 'change-me',
 }
 
 const authHeader = {
